Ship the two health probes instead of leaving them as examples
Echo and Reopen are the operations an application copies into its cell
for health checks. bc5 and HEY have both copied them, and copies drift.
They move into the server gem as health.echo and health.reopen, and a
cell serves them only if it requires hot_cell/health_operations.
