+++ examples/operations/health.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+# The health probes ship with the server gem; the battery uses them like an application would.
+require "hot_cell/health_operations"
